+++ src/full/Agda/Syntax/Concrete/Attribute.hs
@@ -29,6 +29,7 @@ data Attribute
   | QuantityAttribute  Quantity
   | TacticAttribute (Ranged Expr)
   | CohesionAttribute Cohesion
+  | PolarityAttribute PolarityModality
   | LockAttribute      Lock
   deriving (Show)
 
@@ -37,6 +38,7 @@ instance HasRange Attribute where
     RelevanceAttribute r -> getRange r
     QuantityAttribute q  -> getRange q
     CohesionAttribute c  -> getRange c
+    PolarityAttribute p  -> getRange p
     TacticAttribute e    -> getRange e
     LockAttribute l      -> NoRange
 
@@ -45,6 +47,7 @@ instance SetRange Attribute where
     RelevanceAttribute a -> RelevanceAttribute $ setRange r a
     QuantityAttribute q  -> QuantityAttribute  $ setRange r q
     CohesionAttribute c  -> CohesionAttribute  $ setRange r c
+    PolarityAttribute p  -> PolarityAttribute  $ setRange r p
     TacticAttribute e    -> TacticAttribute e  -- -- $ setRange r e -- SetRange Expr not yet implemented
     LockAttribute l      -> LockAttribute l
 
@@ -53,12 +56,13 @@ instance KillRange Attribute where
     RelevanceAttribute a -> RelevanceAttribute $ killRange a
     QuantityAttribute q  -> QuantityAttribute  $ killRange q
     CohesionAttribute c  -> CohesionAttribute  $ killRange c
+    PolarityAttribute p  -> PolarityAttribute  $ killRange p
     TacticAttribute e    -> TacticAttribute    $ killRange e
     LockAttribute l      -> LockAttribute l
 
 -- | (Conjunctive constraint.)
 
-type LensAttribute a = (LensRelevance a, LensQuantity a, LensCohesion a, LensLock a)
+type LensAttribute a = (LensRelevance a, LensQuantity a, LensCohesion a, LensModalPolarity a, LensLock a)
 
 -- | Modifiers for 'Relevance'.
 
@@ -105,6 +109,16 @@ cohesionAttributeTable =
 
 type Attributes = [(Attribute, Range, String)]
 
+-- | Modifiers for 'Polarity'.
+
+polarityAttributeTable :: [(String, PolarityModality)]
+polarityAttributeTable =
+  [ ("unused" , withStandardLock UnusedPolarity)
+  , ("++" , withStandardLock StrictlyPositive)
+  , ("+" , withStandardLock Positive)
+  , ("-" , withStandardLock Negative)
+  , ("mixed" , withStandardLock MixedPolarity)]
+
 -- | Modifiers for 'Quantity'.
 
 lockAttributeTable :: [(String, Lock)]
@@ -122,6 +136,7 @@ attributesMap = Map.fromListWith __IMPOSSIBLE__ $ concat
   [ map (second RelevanceAttribute) relevanceAttributeTable
   , map (second QuantityAttribute)  quantityAttributeTable
   , map (second CohesionAttribute)  cohesionAttributeTable
+  , map (second PolarityAttribute)  polarityAttributeTable
   , map (second LockAttribute)      lockAttributeTable
   ]
 
@@ -144,6 +159,7 @@ setAttribute = \case
   RelevanceAttribute r -> setRelevance r
   QuantityAttribute  q -> setQuantity  q
   CohesionAttribute  c -> setCohesion  c
+  PolarityAttribute  p -> setModalPolarity p
   LockAttribute      l -> setLock      l
   TacticAttribute t    -> id
 
@@ -180,6 +196,13 @@ setPristineCohesion c a
   | getCohesion a == defaultCohesion = Just $ setCohesion c a
   | otherwise = Nothing
 
+-- | Setting 'ModalPolarity' if unset.
+
+setPristinePolarity :: (LensModalPolarity a) => PolarityModality -> a -> Maybe a
+setPristinePolarity c a
+  | getModalPolarity a == defaultPolarity = Just $ setModalPolarity c a
+  | otherwise = Nothing
+
 -- | Setting 'Lock' if unset.
 
 setPristineLock :: (LensLock a) => Lock -> a -> Maybe a
@@ -194,6 +217,7 @@ setPristineAttribute = \case
   RelevanceAttribute r -> setPristineRelevance r
   QuantityAttribute  q -> setPristineQuantity  q
   CohesionAttribute  c -> setPristineCohesion  c
+  PolarityAttribute  p -> setPristinePolarity  p
   LockAttribute      l -> setPristineLock      l
   TacticAttribute{}    -> Just