+++ src/full/Agda/TypeChecking/Monad/Base.hs
@@ -1582,6 +1582,10 @@ instance LensQuantity MetaInfo where
 instance LensRelevance MetaInfo where
   mapRelevance f = mapModality (mapRelevance f)
 
+instance LensModalPolarity MetaInfo where
+  getModalPolarity   = getModalPolarity . getModality
+  mapModalPolarity f = mapModality (mapModalPolarity f)
+
 -- | Name suggestion for meta variable.  Empty string means no suggestion.
 type MetaNameSuggestion = String
 
@@ -1648,6 +1652,9 @@ instance LensRelevance RemoteMetaVariable where
 
 instance LensQuantity RemoteMetaVariable where
   mapQuantity f = mapModality (mapQuantity f)
+
+instance LensModalPolarity RemoteMetaVariable where
+  mapModalPolarity f = mapModality (mapModalPolarity f)
 
 normalMetaPriority :: MetaPriority
 normalMetaPriority = MetaPriority 0
@@ -2114,6 +2121,7 @@ instance LensArgInfo Definition where
 instance LensModality  Definition where
 instance LensQuantity  Definition where
 instance LensRelevance Definition where
+instance LensModalPolarity Definition where
 
 data NumGeneralizableArgs
   = NoGeneralizableArgs
@@ -3971,6 +3979,7 @@ currentModality = do
   q <- viewTC eQuantity
   return Modality
     { modRelevance = r
+    , modPolarity  = defaultPolarity
     , modQuantity  = q
     , modCohesion  = unitCohesion
     }
@@ -4532,6 +4541,8 @@ data TypeError
             -- ^ Wrong user-given quantity annotation in lambda.
         | WrongCohesionInLambda
             -- ^ Wrong user-given cohesion annotation in lambda.
+        | WrongPolarityInLambda
+            -- ^ Wrong user-given polarity annotation in lambda.
         | QuantityMismatch Quantity Quantity
             -- ^ The given quantity does not correspond to the expected quantity.
         | HidingMismatch Hiding Hiding
@@ -4564,6 +4575,7 @@ data TypeError
             --   @SizeUniv@, which is not allowed.
         | SplitOnIrrelevant (Dom Type)
         | SplitOnUnusableCohesion (Dom Type)
+        | SplitOnUnusablePolarity (Dom Type)
         -- UNUSED: -- | SplitOnErased (Dom Type)
         | SplitOnNonVariable Term Type
         | SplitOnNonEtaRecord QName
@@ -4576,6 +4588,7 @@ data TypeError
         | VariableIsIrrelevant Name
         | VariableIsErased Name
         | VariableIsOfUnusableCohesion Name Cohesion
+        | VariableIsOfUnusablePolarity Name PolarityModality
         | UnequalLevel Comparison Level Level
         | UnequalTerms Comparison Term Term CompareAs
         | UnequalTypes Comparison Type Type
@@ -4586,6 +4599,8 @@ data TypeError
             -- ^ The two function types have different relevance.
         | UnequalCohesion Comparison Term Term
             -- ^ The two function types have different cohesion.
+        | UnequalPolarity Comparison Term Term
+            -- ^ The two function types have different polarity.
         | UnequalFiniteness Comparison Term Term
             -- ^ One of the function types has a finite domain (i.e. is a @Partia@l@) and the other isonot.
         | UnequalHiding Term Term
@@ -4702,6 +4717,7 @@ data TypeError
         | GeneralizedVarInLetOpenedModule A.QName
         | MultipleFixityDecls [(C.Name, [Fixity'])]
         | MultiplePolarityPragmas [C.Name]
+        | ExplicitPolarityVsPragma QName
     -- Concrete to Abstract errors
         | NotAModuleExpr C.Expr
             -- ^ The expr was used in the right hand side of an implicit module