+++ src/full/Agda/TypeChecking/Monad/Signature.hs
@@ -182,7 +182,10 @@ addConstant q d = do
   setMutualBlock i q
   where
     new +++ old = new { defDisplay = defDisplay new ++ defDisplay old
-                      , defInstance = defInstance new `mplus` defInstance old }
+                      , defInstance = defInstance new `mplus` defInstance old
+                      , defArgOccurrences = defArgOccurrences old -- Lucas 2022-12-01: for now this works, but not very sound
+                      , defPolarity       = defPolarity old       -- same here
+                      }
 
 -- | A combination of 'addConstant' and 'defaultDefn'. The 'Language'
 -- does not need to be supplied.
@@ -526,7 +529,9 @@ applySection' new ptel old ts ScopeCopyInfo{ renNames = rd, renModules = rm } = 
       -- definition we get will have signature μ \ Δ → B.  This is only valid
       -- for pure modality systems though.
       let ai = defArgInfo def
-          m = unitModality { modCohesion = getCohesion ai }
+          m = unitModality { modCohesion = getCohesion ai
+                           , modPolarity = getModalPolarity ai
+                           }
       localTC (over eContext (map (mapModality (m `inverseComposeModality`)))) $
         copyDef' ts' np def
       where