+++ src/full/Agda/TypeChecking/Pretty.hs
@@ -352,10 +352,17 @@ instance PrettyTCM Quantity where
 instance PrettyTCM Erased where
   prettyTCM = pretty; {-# INLINE prettyTCM #-}
 
+instance PrettyTCM ModalPolarity where
+  prettyTCM = pretty
+
+instance PrettyTCM PolarityModality where
+  prettyTCM = pretty
+
 instance PrettyTCM Modality where
   prettyTCM mod = hsep
     [ prettyTCM (getQuantity mod)
     , prettyTCM (getRelevance mod)
+    , prettyTCM (getModalPolarity mod)
     ]
 {-# SPECIALIZE prettyTCM :: Modality -> TCM Doc #-}